avformat/mov: Removed 1-second delay of packets when streams are not interleaved
2aaf95a210f2d05f11a9fc190760d30304469863 changed the logic for choosing the next sample to be based on sample position if two samples were within 1 second of each other. This was done to reduce seeking in the case of a slower network protocol.

However, if the streams are not interleaved, packets from streams other than the first stream will be delayed by 1 second. This delay may be OK if the application buffers 1+ seconds of data; however, it is not OK to assume that.

A poorly interleaved file is simply not suitable for streaming. FFmpeg should not use hacks to work around poor inputs.