avcodec/videotoolbox: Reconfigure the decoder if needed after encountering any H.264/HEVC parameter set change
The format description should be updated if an H.264/HEVC parameter set change is encountered in the stream. Before this change, the update was only done for H.264 SPS profile changes. After this change, the update will be done for any H.264 SPS/PPS change and for any HEVC VPS/SPS/PPS change.

In order to prevent unnecessary restarts of the `VTDecompressionSession`, we call `VTDecompressionSessionCanAcceptFormatDescription` to see if the new parameter set(s) can be accommodated. (This technique is recommended in “WWDC 2014 Session 513: Direct Access to Video Encoding and Decoding”.)