avcodec/videotoolbox: Pass all H.264 parameter sets to VideoToolbox
There can be multiple SPS/PPS in the H.264 stream. A parameter set can be referenced in a slice by its integer identifier.

Before this change, we were only passing one SPS and one PPS to VideoToolbox, which resulted in incorrect decoding of videos that have multiple SPS/PPS. After this change, we pass all the parameter sets to VideoToolbox similar to the `ff_videotoolbox_hvcc_extradata_create` function.