avcodec/decode: Guess audio PTS by sample count
As an example of why we need to guess, consider the AAC codec. The AAC decoder requires a packet to contain exactly 1024 samples. If a format does not respect this requirement, then the frame PTS should be different from the packet PTS. Therefore, we cannot simply use the packet PTS as the frame PTS in this case.