avformat/avformat: Add `sub_charenc` field to `AVFormatContext`.
When decoding text subtitle packets, the decoder needs to know the character encoding of the text. The demuxer may already know the character encoding if the format specification only allows a single character encoding or if the format explicitly specifies a character encoding. Therefore, we should expose this information to the user to avoid unnecessary guessing in these scenarios.

The field is named `sub_charenc` to match the corresponding field in `AVCodecContext`.