avcodec/htmlsubtitles: Do not stop parsing after encountering an empty line.
There may be more text after an empty line, so we should not stop parsing after encountering an empty line.

I am guessing the original code was written in the context of the SRT file format, which separates subtitles using an empty line. However, that logic should live in (a) an SRT-specific module and (b) in the avformat layer. Putting that logic in this relatively generic HTML subtitles parsing code makes this code less reusable.